# LMS_CODEBASE/packages/openapi/ulms-api.yaml
openapi: 3.0.3
info:
  title: ULMS API
  version: 1.7.0
  description: |
    Unisoft Loan Management System — public API (source of truth per PLANNING/05).
    Conventions: problem+json errors (stable `code`, registry docs/errors.md),
    Idempotency-Key on creating POSTs (replays return the original response
    with `Idempotent-Replay: true`, 48h window), list responses use the
    {data, meta} envelope, money is integer minor units, bilingual objects
    {en,bn} render ONE per request.
    Mutating routes are role-gated (403 ULMS-FORBIDDEN) — realm roles per 06 §1;
    the approval act() ladder-level check maps roles per the band table.
    v1.2 (P2): assessments (CIB/score/DBR/collateral), disbursements with dual
    authorization, compliance EOD/classification/calculator.
    v1.2.1: CIB file channel + 24-month repayment track + raw kept only in the
    encrypted object store (06 §5) + approver inbox.
    v1.3: optimistic concurrency (PATCH version ⇒ 409 ULMS-STATE-0002),
    collateral valuation history, application detail incl. workflow state,
    ?officer= list filter, CIB retry×5 + RB-05 alert on outage.
    v1.4 (P3): payments + rails (intent→redirect→HMAC webhook idempotent by
    UTR), schedules/statements (JSON+CSV), settle quotes, reschedules,
    collections (worklist/actions/PTP/field tasks), tax certificates,
    borrower portal projections.
    v1.4.1: ESCALATE approval action, dunning-ladder timers (queue/run/
...
    portfolio reporting read model.
...
servers:
...
security: [{bearerAuth: []}]
paths:
...
  /collections/{loanId}/actions:
    get:
      tags: [collections]
      summary: Collections action history for a loan (03) — the officer's contact trail
      parameters: [{name: loanId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        "200": {description: list envelope, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {type: object, properties: {actionType: {type: string}, outcome: {type: string}, notes: {type: string, nullable: true}, actedAt: {type: string}, actedBy: {type: string}}}}, meta: {$ref: "#/components/schemas/Meta"}}}}}}
  /collections/{loanId}/actions/{actionType}:
    post:
      tags: [collections]
      summary: Dunning-ladder action (CALL|SMS|VISIT|NOTICE) with outcome
      parameters: [{name: loanId, in: path, required: true, schema: {type: string, format: uuid}}, {name: actionType, in: path, required: true, schema: {type: string, enum: [CALL, SMS, VISIT, NOTICE]}}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [outcome], properties: {outcome: {type: string}, notes: {type: string}}}
      responses:
        "200": {description: action, content: {application/json: {schema: {type: object}}}}
    get:
      tags: [collections]
      summary: Action history for a loan (audit R1 — route existed undocumented)
      parameters: [{name: loanId, in: path, required: true, schema: {type: string, format: uuid}}, {name: actionType, in: path, required: true, schema: {type: string, enum: [CALL, SMS, VISIT, NOTICE]}}]
      responses:
        "200": {description: list envelope of past actions, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {type: object}}, meta: {$ref: "#/components/schemas/Meta"}}}}}}
  /collections/{loanId}/ptp:
    post:
      tags: [collections]
      summary: PTP promise (prototype F3 contextual form) — DPD/class snapshot at promise; kept/broken derives from payments
      parameters: [{name: loanId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [promisedAmountMinor, promisedOn, confidence], properties: {promisedAmountMinor: {type: integer}, promisedOn: {type: string, format: date}, confidence: {type: string, enum: [HIGH, MEDIUM, LOW]}, contactName: {type: string}, contactRelation: {type: string}, contactPhone: {type: string, description: masked in every projection (06 §5)}, remark: {type: string}}}
      responses:
        "200": {description: ptp (phone masked), content: {application/json: {schema: {$ref: "#/components/schemas/Ptp"}}}}
  /collections/ptp/{ptpId}/outcome:
    post:
      tags: [collections]
      summary: PTP state machine transition (PENDING→KEPT|BROKEN, terminal)
      parameters: [{name: ptpId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [kept], properties: {kept: {type: boolean}}}
      responses:
        "200": {description: ptp, content: {application/json: {schema: {$ref: "#/components/schemas/Ptp"}}}}
        "409": {$ref: "#/components/responses/Problem"}
  /collections/ptp/calendar:
    get:
      tags: [collections]
      summary: PTP calendar (promises in a date range)
      parameters: [{name: from, in: query, required: true, schema: {type: string, format: date}}, {name: to, in: query, required: true, schema: {type: string, format: date}}]
      responses:
        "200": {description: list envelope, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {$ref: "#/components/schemas/Ptp"}}, meta: {$ref: "#/components/schemas/Meta"}}}}}}
  /collections/field-tasks:
    post:
      tags: [collections]
      summary: Assign a field-visit task (Expo sync contract, 03)
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [loanId, assignedTo, dueOn], properties: {loanId: {type: string, format: uuid}, assignedTo: {type: string}, dueOn: {type: string, format: date}}}
      responses:
        "200": {description: task, content: {application/json: {schema: {$ref: "#/components/schemas/FieldTask"}}}}
    get:
      tags: [collections]
      summary: Open field tasks (due-date order)
      responses:
        "200": {description: list envelope, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {$ref: "#/components/schemas/FieldTask"}}, meta: {$ref: "#/components/schemas/Meta"}}}}}}
  /collections/field-tasks/{taskId}/complete:
    post:
      tags: [collections]
      summary: Complete from mobile — server-wins status, append-only evidence (replays suppressed but evidence kept, 03)
      parameters: [{name: taskId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        content:
          application/json:
            schema: {type: object, properties: {evidence: {type: string}}}
      responses:
        "200": {description: task with evidence log, content: {application/json: {schema: {$ref: "#/components/schemas/FieldTask"}}}}
...
  /collections/dunning/{actionId}/resolve:
    post:
      tags: [collections]
      summary: Officer acts a queued dunning step — outcome recorded, timer cleared
      parameters: [{name: actionId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [outcome], properties: {outcome: {type: string}, notes: {type: string}}}
      responses:
        "200": {description: action, content: {application/json: {schema: {$ref: "#/components/schemas/DunningStep"}}}}
...
  /collections/{loanId}/legal-case:
    post:
      tags: [collections]
      summary: File a legal case (03 recovery & legal case tracking)
      parameters: [{name: loanId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [claimMinor], properties: {court: {type: string}, filedOn: {type: string, format: date}, claimMinor: {type: integer}, lawyer: {type: string}, notes: {type: string}}}
      responses:
        "200": {description: case, content: {application/json: {schema: {$ref: "#/components/schemas/LegalCase"}}}}
    get:
      tags: [collections]
      summary: Legal cases for a loan
      parameters: [{name: loanId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        "200": {description: list envelope, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {$ref: "#/components/schemas/LegalCase"}}, meta: {$ref: "#/components/schemas/Meta"}}}}}}
  /collections/legal-case/{caseId}/status:
    post:
      tags: [collections]
      summary: Update case status (FILED→ONGOING→WON|LOST|WITHDRAWN)
      parameters: [{name: caseId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [status], properties: {status: {type: string, enum: [FILED, ONGOING, WON, LOST, WITHDRAWN]}}}
      responses:
        "200": {description: case, content: {application/json: {schema: {$ref: "#/components/schemas/LegalCase"}}}}
  /collections/{loanId}/waive-interest:
    post:
      tags: [collections]
      summary: Collections resolution adjustment (03 Fineract row) — waive interest via Fineract transaction; mirror-only migrated loans reject 409
      parameters: [{name: loanId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [amountMinor], properties: {amountMinor: {type: integer}}}
      responses:
        "200": {description: txn, content: {application/json: {schema: {type: object, properties: {fineractTxn: {type: integer}}}}}}
        "409": {$ref: "#/components/responses/Problem"}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  responses:
    Problem:
      description: error
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
  schemas:
    Problem:
      type: object
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        code: {type: string, description: "stable code — registry docs/errors.md (ULMS-NOT-FOUND, ULMS-FORBIDDEN, ULMS-VAL-0001, ULMS-STATE-0001)"}
        detail: {type: string}
        requestId: {type: string, format: uuid}
        fields:
          type: array
          items:
            type: object
            properties: {path: {type: string}, code: {type: string}, message: {type: string}, messageBn: {type: string}}
    Meta:
      type: object
      properties:
        requestId: {type: string, format: uuid}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer}
        totalPages: {type: integer}
...
    Ptp:
      type: object
      properties:
        id: {type: string, format: uuid}
        loanId: {type: string, format: uuid}
        promisedAmountMinor: {type: integer}
        promisedOn: {type: string, format: date}
        confidence: {type: string, enum: [HIGH, MEDIUM, LOW]}
        contactName: {type: string, nullable: true}
        contactRelation: {type: string, nullable: true}
        contactPhoneMasked: {type: string, nullable: true}
        remark: {type: string, nullable: true}
        dpdAtPromise: {type: integer}
        classificationAtPromise: {type: string}
        kept: {type: string, enum: [PENDING, KEPT, BROKEN]}
    FieldTask:
      type: object
      properties:
        id: {type: string, format: uuid}
        loanId: {type: string, format: uuid}
        assignedTo: {type: string}
        dueOn: {type: string, format: date}
        status: {type: string, enum: [OPEN, DONE]}
        notes: {type: string, description: append-only JSON-lines evidence log}
        doneAt: {type: string, format: date-time, nullable: true}
    DunningStep:
      type: object
      properties:
        id: {type: string, format: uuid}
        loanId: {type: string, format: uuid}
        actionType: {type: string, enum: [SMS, CALL, VISIT]}
        outcome: {type: string, description: AUTO_QUEUED until resolved}
        notes: {type: string, nullable: true}
        actor: {type: string}
        actedAt: {type: string, format: date-time}
        dueOn: {type: string, format: date, nullable: true, description: null once acted}
...
    LegalCase:
      type: object
      properties:
        id: {type: string, format: uuid}
        loanId: {type: string, format: uuid}
        caseNo: {type: string}
        court: {type: string, nullable: true}
        filedOn: {type: string, format: date, nullable: true}
        status: {type: string, enum: [FILED, ONGOING, WON, LOST, WITHDRAWN]}
        claimMinor: {type: integer}
        lawyer: {type: string, nullable: true}